added .setBPM and .setTimesignature, renamed files for cleared PR
