+++ Gruntfile.js
@@ -84,15 +84,13 @@ module.exports = function(grunt) {
             'eq': 'src/eq',
             'distortion': 'src/distortion',
             'compressor': 'src/compressor',
-            '_looper': 'src/_looper',
             'looper': 'src/looper',
+            'new_looper': 'src/new_looper',
             'soundRecorder': 'src/soundRecorder',
             'signal': 'src/signal',
             'metro': 'src/metro',
             'peakdetect': 'src/peakDetect',
             'gain': 'src/gain',
-            'monosynth' : 'src/monosynth',
-            'polysynth' : 'src/polysynth'
           },
           useStrict: true,
           wrap: {
