--- examples/monosynth_basic/index.html
@@ -1,10 +1,0 @@
-<head>
-  <script language="javascript" type="text/javascript" src="../../lib/p5.js"></script>
-
-  <script language="javascript" type="text/javascript" src="../../lib/addons/p5.dom.js"></script> 
-
-  <script language="javascript" type="text/javascript" src="../../lib/p5.sound.js"></script> 
-
-  <script language="javascript" type="text/javascript" src="sketch.js"></script> 
-
-</head>
