--- examples/polyphonic_synth/index.html
@@ -1,20 +1,0 @@
-<head>
-  <script language="javascript" src="../../lib/p5.js"></script>
-  <!-- uncomment lines below to include extra p5 libraries -->
-	
-  <script language="javascript" src="../../lib/p5.sound.js"></script>
-  <script language="javascript" src="sketch.js"></script>
-  
-  <!-- this line removes any default padding and style. you might only need one of these values set. -->
-  <style> body {padding: 0; margin: 0;} </style>
-</head>
-
-<body>
-
-
-
-
-
-
-
-</body>
